fix 'fail' tests
